Restructure types folder of core package
