+++ packages/core/src/framework/index.react.ts
@@ -1,4 +1,4 @@
-import type { Signal } from '../types/signal.ts';
+import type { Signal } from '../types/signal/index.ts';
 import type { Framework } from './index.ts';
 
 /**
