+++ packages/core/src/types/form/form.ts
@@ -1,15 +1,20 @@
-import type { QRL } from '@qwik.dev/core';
 import type * as v from 'valibot';
-import type { INTERNAL } from '../values.ts';
-import type { InternalObjectStore } from './field.qwik.ts';
-import type {
-  SubmitEventHandler,
-  SubmitHandler,
-  ValidationMode,
-} from './form.ts';
-import type { Schema } from './schema.ts';
-import type { Signal } from './signal.ts';
-import type { DeepPartial } from './utils.ts';
+import type { INTERNAL } from '../../values.ts';
+import type { InternalObjectStore } from '../field/index.ts';
+import type { Schema } from '../schema/index.ts';
+import type { Signal } from '../signal/index.ts';
+import type { DeepPartial, MaybePromise } from '../utils/index.ts';
+
+/**
+ * Validation mode type.
+ */
+export type ValidationMode =
+  | 'initial'
+  | 'touch'
+  | 'input'
+  | 'change'
+  | 'blur'
+  | 'submit';
 
 /**
  * Form config interface.
@@ -58,7 +63,7 @@ export interface InternalFormStore<TSchema extends Schema = Schema>
   /**
    * The parse function of the form.
    */
-  parse: QRL<(input: unknown) => Promise<v.SafeParseResult<TSchema>>>;
+  parse: (input: unknown) => Promise<v.SafeParseResult<TSchema>>;
 
   /**
    * The submitting state of the form.
@@ -86,4 +91,17 @@ export interface BaseFormStore<TSchema extends Schema = Schema> {
   readonly [INTERNAL]: InternalFormStore<TSchema>;
 }
 
-export { ValidationMode, SubmitHandler, SubmitEventHandler };
+/**
+ * Submit handler type.
+ */
+export type SubmitHandler<TSchema extends Schema> = (
+  output: v.InferOutput<TSchema>
+) => MaybePromise<unknown>;
+
+/**
+ * Submit event handler type.
+ */
+export type SubmitEventHandler<TSchema extends Schema> = (
+  output: v.InferOutput<TSchema>,
+  event: SubmitEvent
+) => MaybePromise<unknown>;
