+++ packages/core/src/types/index.ts
@@ -1,6 +1,6 @@
-export * from './field.ts';
-export * from './form.ts';
-export * from './path.ts';
-export * from './schema.ts';
-export * from './signal.ts';
-export * from './utils.ts';
+export * from './field/index.ts';
+export * from './form/index.ts';
+export * from './path/index.ts';
+export * from './schema/index.ts';
+export * from './signal/index.ts';
+export * from './utils/index.ts';
