+++ packages/core/src/types/path/index.ts
@@ -1,0 +1,8 @@
+export type {
+  Path,
+  PathValue,
+  PathKey,
+  RequiredPath,
+  ValidArrayPath,
+  ValidPath,
+} from './path.ts';
