+++ packages/core/src/types/path/path.test-d.ts
