+++ packages/core/src/types/schema/index.ts
@@ -1,0 +1,1 @@
+export * from './schema.ts';
