+++ packages/core/src/types/signal/index.ts
@@ -1,0 +1,1 @@
+export * from './signal.ts';
