+++ packages/core/src/types/signal/signal.preact.ts
