+++ packages/core/src/types/signal/signal.ts
