+++ packages/core/src/types/signal/signal.vue.ts
